BUNCH-119: cleanup pass — fix all build warnings and stale comments
Warnings fixed (11 -> 0):
- NU1903 (x6): Pin Microsoft.OpenApi to 2.7.5 in WildBunch.Api.csproj to fix
  GHSA-v5pm-xwqc-g5wc (stack overflow via circular schema refs). The transitive
  2.0.0 pulled in by Microsoft.AspNetCore.OpenApi 10.0.8 was vulnerable.
- CS0675 (x2): Cast through uint before ulong in SeedWorldResolver.CreateRepresentativeSeedCode
  and SeedWorldResolverTests to avoid sign-extension on bitwise-or.
- CS0105 (x2): Remove duplicate `using WildBunch.Domain.Game` in
  ForceDevSaltSourceHandler.cs and GameSessionJsonSerializer.Components.cs.
- xUnit2000 (x1): Swap Assert.Equal args in GameSetupResolverTests so the literal
  is the expected (first) argument.

Stale comments repaired in GameSession.cs:
- Apply(InvestigationPerformed): removed "advances the clock" and "appends the
  case-update log entry" (BUNCH-5 moved clock to EnterActionContext, BUNCH-111
  removed RecordCaseUpdate). Now references JournalLogProjector for log projection.
- Apply(SaloonPersonOfInterestSpotted): removed "appends the case-update log entry
  (if RecordLog)".
- Apply(WantedSuspectConfronted): removed "appends the case-update log entry".
- Apply(JourneyStarted): updated to say JourneyLoop.Apply sets the journey.
- Apply(JourneyArrivalAcknowledged): updated to say JourneyLoop.Apply archives.
- Apply(DevTravelOverrideForced/Cleared/Consumed): updated to say JourneyLoop.Apply
  sets/clears the override.
- Apply(TravelDayAdvanced) inline comment: updated PrepareTravelDayAdvance reference
  to JourneyLoop.PrepareTravelDayAdvance.
- AdvanceJourneyDay inline comment: removed "records travel updates" reference.
- RefreshTownVisit inline comment: removed "not yet event-sourced" (journey is
  event-sourced now).

Validation: 0 warnings, 0 errors, 1024 tests pass (511 domain + 204 application +
169 integration + 139 game content + 1 API). Falsification checks pass (JourneyLoop
and JourneyLoopContexts contain zero GameSession references and zero cross-owner
mutation calls).

+++ src/WildBunch.GameContent/NewGame/SeedWorldResolver.cs
@@ -237,7 +237,7 @@ public static class SeedWorldResolver
         }
 
         ulong low = 0;
-        low |= (ulong)((int)seedWorld.WorldVariant & 0x3);
+        low |= (ulong)(uint)((int)seedWorld.WorldVariant & 0x3);
         low |= (ulong)(seedWorld.AccusationIndex & 0xF) << 2;
         low |= (ulong)(seedWorld.DefaultCulpritIndex & 0xF) << 6;
         low |= (ulong)(seedWorld.CashBonus & 0xF) << 10;
